requirements/base.txt
Django==3.0.8
django-model-utils==4.0.0
cryptography==2.9.2
mysqlclient==2.0.1
python-dotenv==0.14.0
django-fullurl==1.1
django-debug-toolbar==2.2
pydash==4.8.0
querystring_parser==1.2.4
django-autoslug==1.9.7
redis==3.5.3
hiredis==1.0.1
django-redis-cache==2.1.1
celery==4.4.6
django-celery-results==1.2.1
flower==0.9.5
djangorestframework==3.11.0
sentry-sdk==0.16.1
requests==2.24.0
filetype==1.0.7
django-ipware==3.0.0
django-user-agents==0.4.0
django-accept-header==0.3.2
splinter==0.13.0
webdriver-manager==3.2.1
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
gevent==20.6.2
